requirements.txt
flask==2.1
flask-sqlalchemy==2.5.1
python-dotenv==0.20.0
gunicorn==20.1.0